Colocate remaining Room types
