+++ packages/liveblocks-core/src/index.ts
@@ -34,6 +34,7 @@ export { shallow } from "./lib/shallow";
 export { LiveList } from "./LiveList";
 export { LiveMap } from "./LiveMap";
 export { LiveObject } from "./LiveObject";
+export type { BroadcastOptions, History, Room } from "./room";
 export type { BaseUserMeta } from "./types/BaseUserMeta";
 export type {
   BroadcastEventClientMsg,
@@ -72,7 +73,6 @@ export type {
 export { OpCode } from "./types/Op";
 export type { Others } from "./types/Others";
 export type { Resolve } from "./types/Resolve";
-export type { BroadcastOptions, History, Room } from "./types/Room";
 export type {
   IdTuple,
   SerializedChild,
